requirements.txt
aiogram==2.21
Jinja2==3.1.2
livereload==2.6.3
python-dotenv==0.20.0